+++ CHANGELOG.md
@@ -118,6 +118,9 @@
 * Renamed "Temperature Timeout" and "SD Status Timeout" in Settings to "Temperature Interval" and "SD Status Interval"
   to better reflect what those values are actually used for.
 * Better behaviour of the settings dialog on mobile devices.
+* Added support for rectangular printer beds with the origin in the center ([#682](https://github.com/foosel/OctoPrint/issues/682) 
+  and [#852](https://github.com/foosel/OctoPrint/pull/852)). Printer profiles now contain a new settings ``volume.origin``
+  which can either be ``lowerleft`` or ``center``. For circular beds only ``center`` is supported.
 
 ### Bug Fixes